move abstract_client to src/server
