+++ cockatrice/CMakeLists.txt
@@ -7,7 +7,6 @@ project(Cockatrice VERSION "${PROJECT_VERSION_MAJOR}.${PROJECT_VERSION_MINOR}.${
 set(cockatrice_SOURCES
     ${VERSION_STRING_CPP}
     # sort by alphabetical order, so that there is no debate about where to add new sources to the list
-    src/client/game_logic/abstract_client.cpp
     src/client/game_logic/key_signals.cpp
     src/client/get_text_with_max.cpp
     src/client/menus/deck_editor/deck_editor_menu.cpp
@@ -226,6 +225,7 @@ set(cockatrice_SOURCES
     src/game/zones/view_zone.cpp
     src/game/zones/view_zone_widget.cpp
     src/main.cpp
+    src/server/abstract_client.cpp
     src/server/chat_view/chat_view.cpp
     src/server/handle_public_servers.cpp
     src/server/local_client.cpp
