+++ cockatrice/src/client/tabs/tab_account.cpp
@@ -1,11 +1,11 @@
 #include "tab_account.h"
 
 #include "../../deck/custom_line_edit.h"
+#include "../../server/abstract_client.h"
 #include "../../server/pending_command.h"
 #include "../../server/user/user_info_box.h"
 #include "../../server/user/user_list_manager.h"
 #include "../../server/user/user_list_widget.h"
-#include "../game_logic/abstract_client.h"
 #include "../sound_engine.h"
 #include "pb/event_add_to_list.pb.h"
 #include "pb/event_remove_from_list.pb.h"
