+++ cockatrice/src/client/tabs/tab_game.cpp
@@ -13,11 +13,11 @@
 #include "../../game/player/player_list_widget.h"
 #include "../../game/zones/card_zone.h"
 #include "../../main.h"
+#include "../../server/abstract_client.h"
 #include "../../server/message_log_widget.h"
 #include "../../server/pending_command.h"
 #include "../../server/user/user_list_manager.h"
 #include "../../settings/cache_settings.h"
-#include "../game_logic/abstract_client.h"
 #include "../network/replay_timeline_widget.h"
 #include "../ui/line_edit_completer.h"
 #include "../ui/phases_toolbar.h"
