+++ cockatrice/src/client/tabs/tab_replays.cpp
@@ -1,9 +1,9 @@
 #include "tab_replays.h"
 
+#include "../../server/abstract_client.h"
 #include "../../server/pending_command.h"
 #include "../../server/remote/remote_replay_list_tree_widget.h"
 #include "../../settings/cache_settings.h"
-#include "../game_logic/abstract_client.h"
 #include "pb/command_replay_delete_match.pb.h"
 #include "pb/command_replay_download.pb.h"
 #include "pb/command_replay_modify_match.pb.h"
