+++ cockatrice/src/client/ui/window_main.h
@@ -20,7 +20,7 @@
 #ifndef WINDOW_H
 #define WINDOW_H
 
-#include "../game_logic/abstract_client.h"
+#include "../../server/abstract_client.h"
 #include "pb/response.pb.h"
 
 #include <QList>
