+++ cockatrice/src/server/abstract_client.cpp
@@ -1,6 +1,5 @@
 #include "abstract_client.h"
 
-#include "../../server/pending_command.h"
 #include "featureset.h"
 #include "get_pb_extension.h"
 #include "pb/commands.pb.h"
@@ -18,6 +17,7 @@
 #include "pb/event_user_left.pb.h"
 #include "pb/event_user_message.pb.h"
 #include "pb/server_message.pb.h"
+#include "pending_command.h"
 
 #include <google/protobuf/descriptor.h>
