+++ cockatrice/src/server/abstract_client.h
