+++ cockatrice/src/server/remote/remote_replay_list_tree_widget.cpp
@@ -1,6 +1,6 @@
 #include "remote_replay_list_tree_widget.h"
 
-#include "../../client/game_logic/abstract_client.h"
+#include "../abstract_client.h"
 #include "../pending_command.h"
 #include "pb/command_replay_list.pb.h"
 #include "pb/response_replay_list.pb.h"
